correct integer division, remove some comments

+++ Agent.scala
@@ -4,7 +4,7 @@ class Agent( val matrix: Array[Array[Int]] = Array.ofDim[Int](2,2),
              val alpha: Double = 0.05,
              val epsilon: Double = 0.05)
 {
-  var selectedAction: Action = if((matrix(0).sum/2)>(matrix(1).sum/2)) c else d
+  var selectedAction: Action = if((matrix(0).sum.toDouble/2)>(matrix(1).sum.toDouble/2)) c else d
   var currentC: Double = 1
   var currentD: Double = 1
 
@@ -13,7 +13,6 @@ class Agent( val matrix: Array[Array[Int]] = Array.ofDim[Int](2,2),
   {
     val row = if (selectedAction == c) 0 else 1
     val col = if (adversaryAction == c) 0 else 1
-//    println(s"get utility: ${matrix(row)(col)}")
     matrix(row)(col)
   }
